requirements.txt
langchain~=0.3.27
langchain-google-genai~=2.1.11
pyyaml~=6.0.2
python-dotenv~=1.1.1
chromadb~=1.1.0